requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
